Upgrade to 0.198

+++ presto-main/src/main/java/com/facebook/presto/execution/resourceGroups/IndexedPriorityQueue.java
@@ -32,7 +32,7 @@ final class IndexedPriorityQueue<E>
 {
     private final Map<E, Entry<E>> index = new HashMap<>();
     private final Set<Entry<E>> queue = new TreeSet<>((entry1, entry2) -> {
-        int priorityComparison = Integer.compare(entry2.getPriority(), entry1.getPriority());
+        int priorityComparison = Long.compare(entry2.getPriority(), entry1.getPriority());
         if (priorityComparison != 0) {
             return priorityComparison;
         }
@@ -42,7 +42,7 @@ final class IndexedPriorityQueue<E>
     private long generation;
 
     @Override
-    public boolean addOrUpdate(E element, int priority)
+    public boolean addOrUpdate(E element, long priority)
     {
         Entry<E> entry = index.get(element);
         if (entry != null) {
@@ -121,10 +121,10 @@ final class IndexedPriorityQueue<E>
     private static final class Entry<E>
     {
         private final E value;
-        private final int priority;
+        private final long priority;
         private final long generation;
 
-        private Entry(E value, int priority, long generation)
+        private Entry(E value, long priority, long generation)
         {
             this.value = requireNonNull(value, "value is null");
             this.priority = priority;
@@ -136,7 +136,7 @@ final class IndexedPriorityQueue<E>
             return value;
         }
 
-        public int getPriority()
+        public long getPriority()
         {
             return priority;
         }
